+++ src/databases/index.ts
@@ -1,3 +1,4 @@
-export { AsyncLocalDatabase } from './async-local-database';
-export { IndexedDBDatabase }  from './indexeddb-database';
-export { MockLocalDatabase }  from './mock-local-database';
+export { AsyncLocalDatabase }   from './async-local-database';
+export { IndexedDBDatabase }    from './indexeddb-database';
+export { LocalStorageDatabase } from './localstorage-database';
+export { MockLocalDatabase }    from './mock-local-database';